backend/requirements.txt
# backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
sqlalchemy-bind-manager
psycopg2-binary # PostgreSQL adapter
pydantic
python-dotenv
httpx
celery
redis
cryptography
PyNacl
